Поправил warning-и при наследовании

+++ Labs/Lab2/task4/Primes/Primes/InputParser.h
@@ -3,4 +3,4 @@
 #include <optional>
 
 bool IsNumber(const std::string& str);
-std::optional<unsigned long> GetUpperBound(int argc, char* argv[]);
+std::optional<unsigned int> GetUpperBound(int argc, char* argv[]);
