+++ Labs/Lab2/task4/Primes/Primes/PrimesGenerator.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <set>
 
-using PrimeSet = std::set<unsigned long>;
+using PrimeSet = std::set<unsigned int>;
 
-PrimeSet GeneratePrimeNumbersSet(const unsigned long upperBound);
+PrimeSet GeneratePrimeNumbersSet(const unsigned int upperBound);
